pyproject.toml
[project]
name = "azure-agent-simple"
version = "0.1.0"
description = "Azure OpenAI trend-analytics demo: Investigator + Writer agents over a small sales warehouse"
requires-python = ">=3.10"
dependencies = [
  "azure-identity>=1.17",
  "openai>=1.40",
  "httpx>=0.27",
  "pandas>=2.1",
  "openpyxl>=3.1",
  "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"
